+++ server/src/nest/accommodations/accommodations.dto.ts
@@ -1,0 +1,5 @@
+import { createZodDto } from 'nestjs-zod';
+import { accommodationCreateBodySchema, accommodationUpdateRequestSchema } from '@trek/shared';
+
+export class AccommodationCreateDto extends createZodDto(accommodationCreateBodySchema) {}
+export class AccommodationUpdateDto extends createZodDto(accommodationUpdateRequestSchema) {}
