+++ server/tests/helpers/rpc-host-deps.ts
@@ -14,7 +14,7 @@ import { ItineraryRpc } from '../../src/nest/assignments/itinerary.rpc';
 import { TripsRpc } from '../../src/nest/trips/trips.rpc';
 import { CostsRpc } from '../../src/nest/budget/costs.rpc';
 import { ReservationsRpc } from '../../src/nest/reservations/reservations.rpc';
-import { AccommodationsRpc } from '../../src/nest/days/accommodations.rpc';
+import { AccommodationsRpc } from '../../src/nest/accommodations/accommodations.rpc';
 import { CollabRpc } from '../../src/nest/collab/collab.rpc';
 import { AtlasRpc } from '../../src/nest/atlas/atlas.rpc';
 import { VacayRpc } from '../../src/nest/vacay/vacay.rpc';
@@ -71,7 +71,7 @@ export function allRpcControllers(): object[] {
     new PlacesRpc(anyService(), anyService(), anyService(), anyService()),
     new DaysRpc(anyService(), anyService(), anyService()),
     new ItineraryRpc(anyService(), anyService(), anyService()),
-    new TripsRpc(anyService(), anyService(), anyService(), anyService(), anyService(), anyService(), anyService()),
+    new TripsRpc(anyService(), anyService(), anyService(), anyService(), anyService(), anyService(), anyService(), anyService()),
     new CostsRpc(anyService(), anyService(), anyService(), anyService()),
     new ReservationsRpc(anyService(), anyService(), anyService()),
     new AccommodationsRpc(anyService(), anyService(), anyService()),
